requirements.txt
python-telegram-bot==21.*
python-dotenv
aiohttp
